[FLINK-9113] [connectors] Fix flushing behavior of bucketing sink for local filesystems
